blog_api/requirements.txt
Django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
pytest>=7.4
pytest-django>=4.5